fix(sql): cast percentile_cont to numeric before round (double-precision type error)
percentile_cont() returns double precision regardless of its numeric input, and
Postgres has no round(double precision, integer) — only round(numeric, int). So
round(percentile_cont(...)/1000, 1) threw "function round(double precision,
integer) does not exist". Wrap the percentile result in ::numeric before the
divide/round in the loading-duration query, in BOTH queries/dashboard.sql and
queries/analytics.sql (same query was added to both).

Type-reviewed: every other round() arg already resolves to numeric (the 100.0*…
percentages, avg/max) — only the two percentile lines per file were affected.

+++ queries/dashboard.sql
@@ -287,8 +287,8 @@ select 'loading · duration' as section,
        case when (props->>'cached')::boolean then 'cache hit' else 'full pipeline' end as kind,
        count(*)                                                                        as n,
        round(avg((props->>'dur_ms')::numeric)/1000, 1)                                 as avg_s,
-       round((percentile_cont(0.5) within group (order by (props->>'dur_ms')::numeric))/1000, 1) as median_s,
-       round((percentile_cont(0.9) within group (order by (props->>'dur_ms')::numeric))/1000, 1) as p90_s,
+       round((percentile_cont(0.5) within group (order by (props->>'dur_ms')::numeric))::numeric/1000, 1) as median_s,
+       round((percentile_cont(0.9) within group (order by (props->>'dur_ms')::numeric))::numeric/1000, 1) as p90_s,
        round(max((props->>'dur_ms')::numeric)/1000, 1)                                 as max_s
 from public.analytics_events
 where event='bio_shown' and props ? 'dur_ms'
